Server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Praveenr-2101",email = "prabhapraveen000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-versatileimagefield (>=3.1,<4.0)",
    "python-magic-bin (>=0.4.14,<0.5.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "drf-extra-fields (>=3.7.0,<4.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
